update: 駒を取れるようにした
