--- src/components/komadai.tsx
@@ -1,5 +1,0 @@
-const Komadai = () => {
-    return <div className="komadai" />;
-}
-
-export default Komadai;
